fix: make sure resource in sign in form is present
If sign in succeeded, it works normally. However, when failed once, all the re-rendered form fails no matter credentials are correct or not. Found out it was because the newly rendered form renders each name of inputs looks like "email" instead of "user[email]", causing devise to fail to authenticate. To solve this, we need to set resource whenever the sign in form needs to be rendered.